Added MPPT data handling
- Added MPPT data handling
- Added data type property to options.json
- Combined can_frame, csv_line and datainput to single file, and refactored accordingly
