--- csv_line.py
@@ -1,11 +1,0 @@
-from datainput import DataInput
-
-class CSVLine(DataInput):
-    def __int__(self, opt):
-        DataInput.__init__(self)
-
-    def __init__(self, opt, serial_str):
-        DataInput.__init__(self)
-
-    def asDatapoints(self):
-        pass
